rename history app in stats
