+++ py_backend/stats/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class HistoryConfig(AppConfig):
+class StatsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'history'
+    name = 'stats'
